+++ src/test/resources/features/usuario.feature
@@ -1,8 +1,6 @@
  # language: pt
 
  Funcionalidade: Gerenciamento de um usuario na Petstore
-
-   Cenario: Criar usuario na loja
    @focus
    Cenario: Criar usuário na loja refletindo o negócio
      Quando crio um usuário
